适配 webview hash 类型，int 改为 int64_t

+++ FTMobileSDK/FTSessionReplay/Recorder/SRWireframe/ViewTreeSnapshot/ViewsRecorder/FTUIActivityIndicatorRecorder.h
@@ -16,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface FTUIActivityIndicatorBuilder : NSObject<FTSRNodeWireframesBuilder>
 @property (nonatomic, strong) FTViewAttributes *attributes;
 @property (nonatomic, assign) CGRect wireframeRect;
-@property (nonatomic, assign) int wireframeID;
+@property (nonatomic, assign) int64_t wireframeID;
 @property (nonatomic, strong) UIColor * backgroundColor;
 @end
